l2

+++ Forum/Forum/Controllers/CommentsController.cs
@@ -1,8 +1,12 @@
 ﻿using AutoMapper;
+using Forum.Auth.Model;
 using Forum.Data.Entities;
 using Forum.Data.Repositories;
 using Forum.DTOs.Comment;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace Forum.Controllers
 {
@@ -14,12 +18,14 @@ namespace Forum.Controllers
         private readonly IMapper _mapper;
         private readonly ITopicsRepository _topicsRepository;
         private readonly ICommentsRepository _commentsRepository;
-        public CommentsController(IPostsRepository postsRepository, IMapper mapper, ITopicsRepository topicsRepository, ICommentsRepository commentsRepository)
+        private readonly IAuthorizationService _authorizationService;
+        public CommentsController(IPostsRepository postsRepository, IMapper mapper, ITopicsRepository topicsRepository, ICommentsRepository commentsRepository, IAuthorizationService authorizationService)
         {
             _topicsRepository = topicsRepository;
             _mapper = mapper;
             _postsRepository = postsRepository;
             _commentsRepository = commentsRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -39,6 +45,7 @@ namespace Forum.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = UserRoles.SimpleUser)]
         public async Task<ActionResult<CommentDTO>> PostAsync(int postId, int topicId, CreateCommentDTO commentDto)
         {
             var topic = await _topicsRepository.Get(topicId);
@@ -47,15 +54,23 @@ namespace Forum.Controllers
             var post = await _postsRepository.GetAsync(topicId, postId);
             if (post == null) return NotFound($"Couldn't find a post with id of {topicId}");
 
-            var comment = _mapper.Map<Comment>(commentDto);
+            var comment = new Comment
+            {
+                Description = commentDto.description,
+                CreationTimeUtc = DateTime.UtcNow,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
+            };
+
+            //var comment = _mapper.Map<Comment>(commentDto);
             comment.PostId = postId;
 
             await _commentsRepository.InsertAsync(comment);
-
-            return Created($"/api/topics/{topicId}/posts/{post.Id}/comment/{comment.Id}", _mapper.Map<CommentDTO>(comment));
+            return Created("", new CreateCommentDTO(comment.Description, comment.CreationTimeUtc));
+            //return Created($"/api/topics/{topicId}/posts/{post.Id}/comment/{comment.Id}", _mapper.Map<CommentDTO>(comment));
         }
 
         [HttpPut("{commentId}")]
+        [Authorize(Roles = UserRoles.SimpleUser)]
         public async Task<ActionResult<CommentDTO>> PostAsync(int postId, int topicId, int commentId, CreateCommentDTO commentDto)
         {
             var topic = await _topicsRepository.Get(topicId);
@@ -67,19 +82,37 @@ namespace Forum.Controllers
             var oldComment = await _commentsRepository.GetAsync(postId, commentId);
             if (oldComment == null) return NotFound();
 
-            _mapper.Map(commentDto, oldComment);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, oldComment, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
+            oldComment.Description = commentDto.description;
+            oldComment.CreationTimeUtc = commentDto.CreationTimeUtc;
+            //_mapper.Map(commentDto, oldComment);
 
             await _commentsRepository.UpdateAsync(oldComment);
 
-            return Ok(_mapper.Map<CommentDTO>(oldComment));
+            return Ok(new CreateCommentDTO(oldComment.Description, oldComment.CreationTimeUtc));
+            //return Ok(_mapper.Map<CommentDTO>(oldComment));
         }
 
         [HttpDelete("{commentId}")]
+        [Authorize(Roles = UserRoles.SimpleUser)]
         public async Task<ActionResult> DeleteAsync(int commentId, int postId)
         {
             var comment = await _commentsRepository.GetAsync(commentId, postId);
             if (comment == null)
                 return NotFound();
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, comment, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             await _commentsRepository.DeleteAsync(comment);
